requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.0
pandas==1.2.5
patsy==0.5.1
Pillow==8.3.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scipy==1.7.0
six==1.16.0
statsmodels==0.12.2
